api/requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
clearbit==0.1.7
click==7.1.2
colorama==0.4.3
cryptography==3.2.1
cssselect==1.1.0
defusedxml==0.6.0
Django==3.0.8
django-cors-headers==3.6.0
django-cors-middleware==1.5.0
django-environ==0.4.5
django-filter==2.3.0
fake-useragent==0.1.11
html5lib==1.1
idna==2.10
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
mysql-connector-python==8.0.21
mysqlclient==2.0.1
packaging==20.4
parse==1.17.0
pluggy==0.13.1
protobuf==3.12.2
py==1.9.0
pycparser==2.20
pyee==7.0.2
pyhunter==1.7
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
pytest==5.4.3
python-decouple==3.3
python3-memcached==1.51
pytz==2020.1
requests==2.24.0
requests-html==0.10.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
tqdm==4.48.2
urllib3==1.25.9
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
